+++ src/routes/+layout.svelte
@@ -25,15 +25,16 @@
   <a href="/crud">CRUD</a> |
   <a href="/super-debug">SuperDebug</a> |
   <a href="/nested">Nested</a> |
-  <a href="/multiple">Multiple</a>
-  <a href="/snapshot">Snapshot</a> |
+  <a href="/multiple">Multiple</a> |
+  <a href="/snapshot">Snapshot</a>
   <br />
   <a href="/dates">Dates</a> |
   <a href="/reset">Reset</a> |
   <a href="/url">URL</a> |
   <a href="/tainted">Tainted</a> |
   <a href="/proxies">Proxies</a> |
-  <a href="/properly-nested">Properly nested</a>
+  <a href="/properly-nested">Properly nested</a> |
+  <a href="/files">Files</a>
 </nav>
 
 <slot />